shared/scripts/rag/requirements.txt
llama-index-core>=0.12.0
llama-index-embeddings-openai>=0.3.0
llama-index-vector-stores-faiss>=0.3.0
llama-index-readers-file>=0.4.0
llama-index-retrievers-bm25>=0.4.0
faiss-cpu>=1.9.0
python-docx>=1.1.0
pymupdf>=1.25.0
